Factorize test_ball_tree_query and test_kd_tree_query
