+++ sklearn/neighbors/tests/test_ball_tree.py
@@ -45,27 +45,6 @@ def brute_force_neighbors(X, Y, k, metric, **kwargs):
     return dist, ind
 
 
-@pytest.mark.parametrize('metric', METRICS)
-@pytest.mark.parametrize('k', (1, 3, 5))
-@pytest.mark.parametrize('dualtree', (True, False))
-@pytest.mark.parametrize('breadth_first', (True, False))
-def test_ball_tree_query(metric, k, dualtree, breadth_first):
-    rng = check_random_state(0)
-    X = rng.random_sample((40, DIMENSION))
-    Y = rng.random_sample((10, DIMENSION))
-
-    kwargs = METRICS[metric]
-
-    bt = BallTree(X, leaf_size=1, metric=metric, **kwargs)
-    dist1, ind1 = bt.query(Y, k, dualtree=dualtree,
-                           breadth_first=breadth_first)
-    dist2, ind2 = brute_force_neighbors(X, Y, k, metric, **kwargs)
-
-    # don't check indices here: if there are any duplicate distances,
-    # the indices may not match.  Distances should not have this problem.
-    assert_array_almost_equal(dist1, dist2)
-
-
 @pytest.mark.parametrize('metric',
                          itertools.chain(BOOLEAN_METRICS, DISCRETE_METRICS))
 def test_ball_tree_query_metrics(metric):
