+++ sklearn/neighbors/tests/test_kd_tree.py
@@ -22,37 +22,6 @@ METRICS = {'euclidean': {},
            'minkowski': dict(p=3)}
 
 
-def brute_force_neighbors(X, Y, k, metric, **kwargs):
-    D = DistanceMetric.get_metric(metric, **kwargs).pairwise(Y, X)
-    ind = np.argsort(D, axis=1)[:, :k]
-    dist = D[np.arange(Y.shape[0])[:, None], ind]
-    return dist, ind
-
-
-def check_neighbors(dualtree, breadth_first, k, metric, X, Y, kwargs):
-    kdt = KDTree(X, leaf_size=1, metric=metric, **kwargs)
-    dist1, ind1 = kdt.query(Y, k, dualtree=dualtree,
-                            breadth_first=breadth_first)
-    dist2, ind2 = brute_force_neighbors(X, Y, k, metric, **kwargs)
-
-    # don't check indices here: if there are any duplicate distances,
-    # the indices may not match.  Distances should not have this problem.
-    assert_array_almost_equal(dist1, dist2)
-
-
-@pytest.mark.parametrize('metric', METRICS)
-@pytest.mark.parametrize('k', (1, 3, 5))
-@pytest.mark.parametrize('dualtree', (True, False))
-@pytest.mark.parametrize('breadth_first', (True, False))
-def test_kd_tree_query(metric, k, dualtree, breadth_first):
-    rng = check_random_state(0)
-    X = rng.random_sample((40, DIMENSION))
-    Y = rng.random_sample((10, DIMENSION))
-
-    kwargs = METRICS[metric]
-    check_neighbors(dualtree, breadth_first, k, metric, X, Y, kwargs)
-
-
 def test_kd_tree_query_radius(n_samples=100, n_features=10):
     rng = check_random_state(0)
     X = 2 * rng.random_sample(size=(n_samples, n_features)) - 1
